Move first-class types in silkrpc namespace Add Filter as silkrpc first-class type
