+++ silkrpc/commands/eth_api.hpp
@@ -28,19 +28,17 @@
 
 #include <silkworm/types/receipt.hpp>
 #include <silkrpc/core/rawdb/accessors.hpp>
-#include <silkrpc/core/types/log.hpp>
-#include <silkrpc/core/types/receipt.hpp>
 #include <silkrpc/croaring/roaring.hh>
 #include <silkrpc/json/types.hpp>
 #include <silkrpc/ethdb/kv/database.hpp>
 #include <silkrpc/ethdb/kv/transaction.hpp>
 #include <silkrpc/ethdb/kv/transaction_database.hpp>
+#include <silkrpc/types/log.hpp>
+#include <silkrpc/types/receipt.hpp>
 
 namespace silkrpc::http { class RequestHandler; }
 
 namespace silkrpc::commands {
-
-using namespace silkrpc::core;
 
 class EthereumRpcApi {
 public:
@@ -55,10 +53,10 @@ private:
     asio::awaitable<void> handle_eth_block_number(const nlohmann::json& request, nlohmann::json& reply);
     asio::awaitable<void> handle_eth_get_logs(const nlohmann::json& request, nlohmann::json& reply);
 
-    asio::awaitable<Roaring> get_topics_bitmap(core::rawdb::DatabaseReader& db_reader, json::FilterTopics& topics, uint64_t start, uint64_t end);
-    asio::awaitable<Roaring> get_addresses_bitmap(core::rawdb::DatabaseReader& db_reader, json::FilterAddresses& addresses, uint64_t start, uint64_t end);
+    asio::awaitable<Roaring> get_topics_bitmap(core::rawdb::DatabaseReader& db_reader, FilterTopics& topics, uint64_t start, uint64_t end);
+    asio::awaitable<Roaring> get_addresses_bitmap(core::rawdb::DatabaseReader& db_reader, FilterAddresses& addresses, uint64_t start, uint64_t end);
     asio::awaitable<Receipts> get_receipts(core::rawdb::DatabaseReader& db_reader, uint64_t number, evmc::bytes32 hash);
-    std::vector<Log> filter_logs(std::vector<Log>& logs, const json::Filter& filter);
+    std::vector<Log> filter_logs(std::vector<Log>& logs, const Filter& filter);
 
     std::unique_ptr<ethdb::kv::Database>& database_;
